# storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TasteAsia RESTful API",
        "description": "RESTful API Documentation for TasteAsia. Explore our endpoints and discover how to interact with our delicious data!",
        "contact": {
            "email": "gabrielaswinta@cosrsivalab.space"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://tasteasia.test",
            "description": "TasteAsia API Server"
        },
        {
            "url": "https://tasteasia.test/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "${STAGING_URL}/api/v1",
            "description": "STAGING API Server"
        },
        {
            "url": "https://tasteasia.test/api/v1",
            "description": "PRODUCTION API Server"
        }
    ],
    "paths": {
...
        "/m/register-send-otp": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Send OTP to a phone number for registration",
                "operationId": "registerSendOTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+15017122661"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OTP sent successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to send OTP",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/m/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Register a new user",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "User registration details (multipart/form-data)",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully registered!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/LoginResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Registration failed due to an error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "description": "Request body for user registration",
                "required": [
                    "name",
                    "phone",
                    "otp",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "User's full name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "phone": {
                        "description": "User's phone number (e.g., +6588837454)",
                        "type": "string",
                        "maxLength": 20
                    },
                    "email": {
                        "description": "User's unique email address",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "otp": {
                        "description": "One-time password verification code",
                        "type": "integer",
                        "example": "123456"
                    },
                    "password": {
                        "description": "User's password (will be hashed)",
                        "type": "string",
                        "format": "password",
                        "minLength": 8
                    },
                    "password_confirmation": {
                        "description": "Password confirmation to ensure accuracy",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "title": "Login Response",
                "description": "Response structure for successful login",
                "required": [
                    "id",
                    "access_token",
                    "access_token_expires_at",
                    "refresh_token",
                    "refresh_token_expires_at"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "access_token": {
                        "description": "Access Token",
                        "type": "string"
                    },
                    "access_token_expires_at": {
                        "description": "Timestamp of access token expiration",
                        "type": "integer"
                    },
                    "refresh_token": {
                        "description": "Refresh Token",
                        "type": "string"
                    },
                    "refresh_token_expires_at": {
                        "description": "Timestamp of refresh token expiration",
                        "type": "integer"
                    },
                    "customer_id": {
                        "description": "Customer ID (if the user is a customer)",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "role": {
                        "description": "Customer role (if the user is a customer)",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "permissions": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "description": "Permission names (if the user is a customer)",
                            "type": "string"
                        }
                    },
                    "credit_balance": {
                        "description": "Credit balance (if the user has the 'have credit' permission)",
                        "type": [
                            "number",
                            "null"
                        ],
                        "format": "float"
                    },
                    "cash_balance": {
                        "description": "Cash balance (if the user has the 'have cash' permission)",
                        "type": [
                            "number",
                            "null"
                        ],
                        "format": "float"
                    },
                    "cash_split": {
                        "description": "Cash split (if the user has the 'have cash' permission)",
                        "type": [
                            "number",
                            "null"
                        ],
                        "format": "float"
                    },
                    "credit_split": {
                        "description": "Credit split (if the user has the 'have cash' permission)",
                        "type": [
                            "number",
                            "null"
                        ],
                        "format": "float"
                    },
                    "is_first_login": {
                        "description": "to identify if the user is first login or not (using 1 or 0)",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "required": [
                            "code",
                            "message",
                            "status"
                        ],
                        "properties": {
                            "code": {
                                "description": "Error code from the ApiErrorCode enum",
                                "type": "string",
                                "enum": [
                                    "SERVER_ERROR",
                                    "UNAUTHORIZED",
                                    "BAD_REQUEST",
                                    "INVALID_TOKEN",
                                    "TOKEN_EXPIRED",
                                    "TOKEN_MISSING",
                                    "TOKEN_STOLEN",
                                    "TOKEN_TYPE",
                                    "USER_NOT_FOUND",
                                    "REGISTRATION_ERROR",
                                    "VALIDATION_ERROR"
                                ]
                            },
                            "message": {
                                "description": "Error message",
                                "type": "string"
                            },
                            "status": {
                                "description": "HTTP status code",
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer Token": {
                "type": "http",
                "description": "Enter token in format 'Bearer {token}'. Token is a 64-character random string.",
                "scheme": "bearer"
            },
            "Access Token": {
                "type": "apiKey",
                "description": "Sanctum Access Token (a random string)",
                "name": "X-Taste-Asia-Access-Token",
                "in": "header"
            },
            "Refresh Token": {
                "type": "apiKey",
                "description": "Sanctum Refresh Token (a random string)",
                "name": "X-Taste-Asia-Refresh-Token",
                "in": "header"
            }
        }
    },
...
}